requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2025.4
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.7
django-appconf==1.1.0
django-crispy-forms==2.4
django-currentuser==0.9.0
django-filter==25.1
django-import-export==4.3.7
django-minify-html==1.12.0
django-password-eye==1.1.4
django-pivot==1.10.0
django-select2==8.4.0
django-tables2==2.7.5
django-tinymce==4.1.0
et_xmlfile==2.0.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
minify_html==0.16.4
multidict==6.4.3
openpyxl==3.1.5
packaging==25.0
pillow==11.1.0
Pygments==2.19.1
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.3.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.4.0
